[IMP] Improved style as per OCA guidelines
