--- account_advanced_reconcile/easy_reconcile.py
@@ -1,37 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Guewen Baconnier
-#    Copyright 2012 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-from openerp import models, api
-
-
-class account_easy_reconcile_method(models.Model):
-
-    _inherit = 'account.easy.reconcile.method'
-
-    @api.model
-    def _get_all_rec_method(self):
-        methods = super(account_easy_reconcile_method, self).\
-            _get_all_rec_method()
-        methods += [
-            ('easy.reconcile.advanced.ref',
-             'Advanced. Partner and Ref.'),
-        ]
-        return methods
