+++ account_advanced_reconcile/models/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 # © 2012-2014 Camptocamp SA - Guewen Baconnier
 
-from . import models
-from . import wizards
+from . import easy_reconcile
+from . import base_advanced_reconciliation
