+++ account_advanced_reconcile/wizards/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 # © 2012-2014 Camptocamp SA - Guewen Baconnier
 
-from . import models
-from . import wizards
+from . import advanced_reconciliation
